mqtt_to_influxdb_poller/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
importlib-metadata==0.23
influxdb==5.2.3
more-itertools==7.2.0
packaging==19.2
paho-mqtt==1.5.0
pkg-resources==0.0.0
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.4
pytest==5.2.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
